+++ extensions/google/test-api.ts
@@ -1,5 +1,6 @@
 export { buildGoogleGeminiCliBackend } from "./cli-backend.js";
 export { buildGoogleImageGenerationProvider } from "./image-generation-provider.js";
 export { buildGoogleMusicGenerationProvider } from "./music-generation-provider.js";
+export { buildGoogleSpeechProvider } from "./speech-provider.js";
 export { googleMediaUnderstandingProvider } from "./media-understanding-provider.js";
 export { buildGoogleVideoGenerationProvider } from "./video-generation-provider.js";